refactor: peel file io into lingtai sdk
